verifying request method is delete.

+++ pages/api/contacts/[ref].js
@@ -1,6 +1,8 @@
 import { deleteContact } from "../../../services/contacts";
 
 export default async function handler(req, res) {
-  await deleteContact(req.query.ref);
-  res.json({ OK: true })
+  if (req.method === 'DELETE') {
+    await deleteContact(req.query.ref);
+    res.json({ OK: true })
+  }
 }
